fixed some warnings and one signed/unsigned bug
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@291 3db68df2-c116-0410-a063-a993310a9797

+++ source/gui/CGUIScrollBarVertical.cpp
@@ -22,7 +22,7 @@ void CGUIScrollBarVertical::SetPosFromMousePos(const CPos &mouse)
 	if (!GetStyle())
 		return;
 
-	m_Pos = (m_PosWhenPressed + (float)m_ScrollRange*((float)mouse.y-m_BarPressedAtPos.y)/(m_Length-GetStyle()->m_Width*2));
+	m_Pos = (int)(m_PosWhenPressed + (float)m_ScrollRange*((float)mouse.y-m_BarPressedAtPos.y)/(m_Length-GetStyle()->m_Width*2));
 }
 
 void CGUIScrollBarVertical::Draw()
